abstract voiceinput to a separate component
