+++ app/events/page.tsx
@@ -7,46 +7,43 @@ import { EventRole, EventVisibility } from "@/lib/types";
 export const dynamic = "force-dynamic";
 
 export default async function EventsPage() {
-
   async function fetchEvents() {
-
     const supabase = await createSupabaseClient();
-  
+
     try {
       // Check if user is authenticated
       const {
         data: { user },
       } = await supabase.auth.getUser();
-      
+
       if (!user) {
         console.error("User not authenticated");
         return [];
       }
-  
+
       const userId = user.id;
-  
+
       // Get user's roles in all events
       const { data: userRoles, error: rolesError } = await supabase
         .from("user_event_roles")
         .select("event_id, role")
         .eq("user_id", userId);
-  
+
       if (rolesError) throw rolesError;
-  
+
       // Check if user is an admin in any event
-      const { data } = await supabase
-        .rpc('is_admin_user', { user_id: userId });
+      const { data } = await supabase.rpc("is_admin_user", { user_id: userId });
 
       const isAdmin = data || false;
-      
+
       // Get event IDs where user is an organizer
       const organizerEventIds = userRoles
-        .filter(role => role.role === EventRole.Organizer)
-        .map(role => role.event_id);
-  
+        .filter((role) => role.role === EventRole.Organizer)
+        .map((role) => role.event_id);
+
       // Build query for events based on visibility and user role
       let query = supabase.from("events").select("*");
-  
+
       if (isAdmin) {
         // Admins can see all events (public, private, draft, demo, test)
         // No additional filter needed
@@ -56,30 +53,45 @@ export default async function EventsPage() {
         // - Draft events where user is an organizer
         // - No demo or test events
         query = query.or(
-          `visibility.in.(${[EventVisibility.Public, EventVisibility.Private].join(",")}),` + 
-          `and(visibility.eq.${EventVisibility.Draft},event_id.in.(${organizerEventIds.join(",")}))`
+          `visibility.in.(${[
+            EventVisibility.Public,
+            EventVisibility.Private,
+          ].join(",")}),` +
+            `and(visibility.eq.${
+              EventVisibility.Draft
+            },event_id.in.(${organizerEventIds.join(",")}))`
         );
       } else {
         // Regular users only see public and private events
-        query = query.in("visibility", [EventVisibility.Public, EventVisibility.Private]);
+        query = query.in("visibility", [
+          EventVisibility.Public,
+          EventVisibility.Private,
+        ]);
       }
-  
+
       const { data: events, error: eventsError } = await query;
-  
+
       if (eventsError) throw eventsError;
-  
+
       // Include user role with each event
-      const eventsWithRole = events.map(event => {
-        const userRole = userRoles.find(role => role.event_id === event.event_id)?.role || null;
-        return { ...event, userRole };
-      });
-  
+      const eventsWithRole = events
+        .map((event) => {
+          const userRole =
+            userRoles.find((role) => role.event_id === event.event_id)?.role ||
+            null;
+          return { ...event, userRole };
+        })
+        .sort(
+          (a, b) =>
+            new Date(b.event_date).getTime() - new Date(a.event_date).getTime()
+        );
+
       return eventsWithRole;
     } catch (error) {
       console.error("Error fetching events:", error);
       return [];
     }
-  } 
+  }
 
   const events = await fetchEvents();
